refactor(daemon): move runner.ts into lib/daemon/
Second hub-drain slice: the routines job executor already belongs next to
daemon.ts. No behavior change — import retargets only, no hub re-export
shim.
